@@ -1,38 +1,47 @@
 <script lang="ts">
+	import { run, preventDefault } from 'svelte/legacy';
+
 	import { confetti } from '@neoconfetti/svelte';
 	import { enhance } from '$app/forms';
 	import type { PageData, ActionData } from './$types';
 	import { reduced_motion } from './reduced-motion';
 
-	export let data: PageData;
 
-	export let form: ActionData;
+	interface Props {
+		data: PageData;
+		form: ActionData;
+	}
+
+	let { data, form = $bindable() }: Props = $props();
 
 	/** Whether or not the user has won */
-	$: won = data.answers.at(-1) === 'xxxxx';
+	let won = $derived(data.answers.at(-1) === 'xxxxx');
 
 	/** The index of the current guess */
-	$: i = won ? -1 : data.answers.length;
+	let i = $derived(won ? -1 : data.answers.length);
 
 	/** The current guess */
-	$: currentGuess = data.guesses[i] || '';
+	let currentGuess;
+	run(() => {
+		currentGuess = data.guesses[i] || '';
+	});
 
 	/** Whether the current guess can be submitted */
-	$: submittable = currentGuess.length === 5;
+	let submittable = $derived(currentGuess.length === 5);
 
 	/**
 	 * A map of classnames for all letters that have been guessed,
 	 * used for styling the keyboard
 	 */
-	let classnames: Record<string, 'exact' | 'close' | 'missing'>;
+	let classnames: Record<string, 'exact' | 'close' | 'missing'> = $state();
 
 	/**
 	 * A map of descriptions for all letters that have been guessed,
 	 * used for adding text for assistive technology (e.g. screen readers)
 	 */
-	let description: Record<string, string>;
+	let description: Record<string, string> = $state();
 
-	$: {
+	run(() => {
 		classnames = {};
 		description = {};
 
@@ -51,7 +60,7 @@
 				}
 			}
 		});
-	}
+	});
 
 	/**
 	 * Modify the game state without making a trip to the server,
@@ -83,7 +92,7 @@
 	}
 </script>
 
-<svelte:window on:keydown={keydown} />
+<svelte:window onkeydown={keydown} />
 
 <svelte:head>
 	<title>Sverdle</title>
@@ -150,7 +159,7 @@
 				<button data-key="enter" class:selected={submittable} disabled={!submittable}>enter</button>
 
 				<button
-					on:click|preventDefault={update}
+					onclick={preventDefault(update)}
 					data-key="backspace"
 					formaction="?/update"
 					name="key"
@@ -163,7 +172,7 @@
 					<div class="row">
 						{#each row as letter}
 							<button
-								on:click|preventDefault={update}
+								onclick={preventDefault(update)}
 								data-key={letter}
 								class={classnames[letter]}
 								disabled={submittable}
@@ -192,7 +201,7 @@
 			stageHeight: window.innerHeight,
 			colors: ['#ff3e00', '#40b3ff', '#676778']
 		}}
-	/>
+	></div>
 {/if}
 
 <style>
